+++ spec/strategies/history_spec.rb
@@ -8,5 +8,16 @@ describe 'the `history` suggestion strategy' do
     end
   end
 
+  context 'when ZSH_AUTOSUGGEST_HISTORY_IGNORE is set to a pattern' do
+    let(:options) { ['ZSH_AUTOSUGGEST_HISTORY_IGNORE="* bar"'] }
+
+    it 'does not make suggestions that match the pattern' do
+      with_history('ls foo', 'ls bar', 'echo baz') do
+        session.send_string('ls')
+        wait_for { session.content }.to eq('ls foo')
+      end
+    end
+  end
+
   include_examples 'special characters'
 end